Prevent using RoundTripper after it was closed
Signed-off-by: Marius Sincovici <marius.sincovici@nordsec.com>

The inner RoundTripper can be used by multiple requests in the same time.
It can be closed on error or on VPN connection. This 2 operation can overlap
resulting in a crash because requests are using a closed RoundTripper.
To prevent the crash protect access to RoundTripper with a RWLock.
Plus add an atomic.Bool to prevent recreating RoundTripper by 2 concurrent
requests on the same error.