Add wait for iptables (#531)
Signed-off-by: Marius Sincovici <marius.sincovici@nordsec.com>

In some cases iptables are locked and the command fails with Another app is currently holding the xtables lock. Perhaps you want to use the -w option?.
To prevent this add -w into the param list for iptables Add also for how long to wait for the lock to be free, otherwise the daemon gets stuck.