Start moose worker before the tracker
This reason for this change is that there is an issue with reading env
vars on the moose side. Moose worker is reading env vars and moose
tracker is doing some changes in separate threadd leading to invalid
pointer access in worker.
The workaround is to reorder the initialization of moose worker and
tracker.