Introduce meshnet map caching
When meshnet map is requested from the `mesh.Retriever`, instead of
calling API directly, it is called via `CachingRetriever`. Also,
`NotifyingRegistry` was introduced as well to make sure that any changes
to meshnet map made by this device would notify the subscribers about
it. This also means, that application now reacts to these changes in a
same way as it does to NC messages, therefore it is no longer needed to
call networker directly from meshnet map modifying gRPC functions.