Handle meshnet key registration delay
There is a delay on the backed between registering a new
meshnet/NordLynx key and the key being recognized by the server. It may
lead to connection failures, for example when user enables
meshnet(registering a new key) and then immediately tries to connect
usint NordLynx(key is not recognized yet). A sleep was added to ensure a
wait time after new key is registered.