LVPN-9096: Create symlink for sqlite in posttrans script (#1090)
For RPM packets are 2 different scripts that are executed at package install/update: postinstall and posttrans.
They are executed in different times while installing. The posttrans is executed the latest, after the update/install process takes place. postinstall is executed before new package replaces the files of the old one with its own.
Since sqlite symlink is not part of the new package the file is deleted after postinstall is executed. Creating it into posttrans ensures that the file is not deleted.