LVPN-9097: Fix tests failing due to the new version available message
This replaces re.match with re.search when checking the connect command
output.
re.match will try to match starting from the beginning of the stdout.
re.search will try to find the first match in the entire stdout.
In this way, the test will no longer be affected by additional messages
like the one informing the user about a new version available.

Signed-off-by: Daniel Fetti <daniel.fetti@nordsec.com>