LVPN-9736: set app to defaults in meshnet test setup
When a test fails in setup, the teardown will not be performed. In such
cases we need to perform the necessary teardown in the setup function,
otherwise next test will be started in invalid state. Primarily, this
concerns setting the app to default/logging out as otherwise the next test will fail
at the login attempt.

Signed-off-by: Bartosz Oleaczek <bartosz.oleaczek@nordsec.com>