LVPN-9442: resolv.conf based DNS management service detection
resolv.conf DNS management service detection was addedd in order to
ensure that the daemon will use the configuration method supported by
the OS.

If a comment in resolv.conf matches a service, that service will be used
to configure the DNS.

If no comment is recognized, the daemon will read a symlink target. If
the symlink target matches some service, that service will be used to
configure the DNS.

If resolv.conf is not a symlink or symlink target is not recognized,
daemon will attempt to set the DNS using the best available method,
based on a priority list: 1, systemd-resolved 2, direct changes in
resolv.conf.